Modified the Constitutive class (and the derived class Elastoviscoplasticity) to enable its state to be initialized from (or exported to) a binary block that is allocated and owned by LULESH.
Added 3 new members to the LULESH domain struct: a void pointer (domain.cm_state) to the constitutive model state (i.e., the binary block described above), a Tensor2Gen object (domain.cm_vel_grad) to store the velocity gradient computed in CalcKinematicsForElems(), and a double (domain.cm_vol_chng) to store the volume change also computed there.

Changed the constitutive model advance() to now take the velocity gradient and volume change as input, as well as the time step (as before) and the current state.  Upon return, in addition to the returned ConstitutiveData struct, the state block is also updated for use in the next call.  Because they are no longer needed, the setVelocityGradient() and setVolumeChange() constitutive model members have been removed.