Adding curl for systems (MacOS) without wget
Newly cloned CoEVP repos require that both redis and flann
be installed locally. The Makefiles in their respective
directories download tarballs, unpack them and then build the
packages. Previously, wget was used to download the tarballs
and this command isn't present on OSX. Added logic in Makefiles
to choose curl or wget depending on system in use.