Build everything; select execution via cmd line
I don't like #defines so I removed most of the extraneous ones
and simply build everythig (REDIS, FLANN) into the CM library.
After all, we'll be doing A/B tests at some point so we may as
well build everything in now--it will make testing easier.

I also removed the FLANN #defines in LULESH. Both types of adaptive
sampling are comiled in and selectd with command line arguments
(-s for adaptive sampling and -s -f for adaptive sampling with
the FLANN library (vs MTree). No adaptive is again the default
simply because you can't turn off an on default (of that makes
sense).

I also removed -fopenmp from the build because it was producing
multiple runtime memory errors. This will take some more debugging
if we want to use OpenMP.

/cc @milodorr
/cc @junghans