Added counters/timers to adaptive sampler
The AdaptiveSampler::sample method is a good place to sprinkle
timers and counters (as the author of the FSTRACE stuff also
found and I should have noticed earlier). I added the following
logger functions to this method:

 - Time the interpolate function
 - Count interpolation failures
 - Time fine-scale model evaluation on failure
 - Time insert of fine-scale value into databse
 - Count interpolation sucesses
 - Time "second?" interpolation on success (@milodorr)

Yet to do:

 - Write Python scripts to analyze collected data
 - Merge the logger stuff back up in to the master (ugh!)

FYI, I ran this on Darwin (-s, four nodes via MPI) with a fifth
node for the logging databse (in addition to the REDIS processes
that the adaptive sampler code runs). The logger hit the databse
at about 2000 writes/second. I think it does slow down the code
execution a bit, but I have a fix for caching counters that should
fix that.