Added lctest for testing libcircle (temporary file)
Makefile in serverize directory will build both protobuf and
libcircle. Makefile in LULESH is still broken. You need to
run protoc once by hand to get started, then make will run
protobuf as needed when the .proto file is updated. Command is:

../serverize/protobuf/bin/protoc --cpp_out=. advance.proto

The lctest.cc file just a playground for libcircle experimentation.
It's easier to get the flow of the code working in there before
integrating it into lulesh. As of this commit it just tests a
simple path of execution--nothing fancy. Since libcircle is NOT
an RPC library, it does not return values from its "work" functions.
We'll need to use the datanase (REDIS) for that.

Finally, as of this commit, lctest workes fine on a single node. If
you try and run multi-node, ala:

mpirun --map-by node -np 2 ./lctest

it will hang. Might be able to get some help form @spakin on this
(he's likely the local expert having done a Go version).

/cc @louisvernon