Libcircle working (including added work)
This libcircle implementation (as ugly as it is) seems to
be working well enough to start timing/scaling tests. The
database must be deleted/restarted for every run (don't know
why code that automates this isn't working). I have also added
some "fake" work to advance so we can measure something besides
overhead (database and libcircle).

I will start running some timing/scaling tests on Darwin. Next
week I will attempt to build and run on Trinity.

Note: NEVER try to merge this branch with the master branch. This is
a messy dead end branch. If we want to merge a libcircle implementation
into master, I'll clean it up (in a new branch) after the Trinity
deliverables are completed.